backend/requirements.txt
aniso8601==2.0.0
blinker==1.4
click==6.7
dateutils==0.6.6
Flask==0.12.2
Flask-Admin==1.5.0
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
marshmallow==2.15.0
peewee==2.10.2
Pillow==5.0.0
pyldap==2.4.45
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
uWSGI==2.0.15
webargs==1.8.1
Werkzeug==0.14.1
wtf-peewee==0.2.6
WTForms==2.1
WeasyPrint==50
requests==2.22.0